pyproject.toml
[tool.poetry]
name = "complex-valued-nn-for-ToyADMOS"
version = "0.1.0"
description = ""
authors = ["nissy-shota <shota_4410@icloud.com>"]
packages = [
    { include = "complexmodules" }
]

[tool.poetry.dependencies]
python = "^3.8"
librosa = "^0.9.2"
PyYAML = "^6.0"
matplotlib = "^3.5.2"
mlflow = "^1.27.0"
slackweb = "^1.0.5"
hydra-core = "^1.2.0"
torchinfo = "^1.7.0"
joblib = "^1.1.0"
python-dotenv = "^0.20.0"
torch = { url = "https://download.pytorch.org/whl/cu116/torch-1.13.0%2Bcu116-cp38-cp38-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu116/torchvision-0.14.0%2Bcu116-cp38-cp38-linux_x86_64.whl"}
timm = "^0.6.7"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.950"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88
max-complexity = 18
ignore = ["E203", "E266", "W503"]

[tool.isort]
profile = "black"
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
